refactor: Remove dead code
We are looping over the arguments set by the parser.  In one case, we
are checking to see if the argument is a group.  I am not seeing us set
groups in the matcher when parsing and we have a debug assert ensuring
arg and group names don't conflict.  This looks like dead code to me.

I am hesitant making a change like this right before release but this is
part of trying to make the code understandable for root causing and
fixing #3197