feat: Add `has_no_matched_args` to `ArgMatches`
This member function provides a simple way to check if no arguments
were matched at all, in v3. The same in v2 could be achieved by
accessing the `args` field in `ArgMatches` and checking if it was
empty. However, in v3 the field was made private in the public API
and for the same reason, this method was added.