fix(help): Improve multiple usage statements
Repairs #1068

Some applications/subcommands have multiple distinct usages that demand
multiple usage statements. This change attempts to untangle various
assumptions about usage strings being single-line such that multi-line
usage strings are better supported while remaining mostly backwards
compatible with user-provided help templates.

A new template tag, `{usage-indented}` is introduced and replaces `{usage}`
in the default help templates. As its name implies, this new tag displays
each usage line within the usage string with indentation. In the default
templates (DEFAULT_TEMPLATE and DEFAULT_NO_ARGS_TEMPLATE),
`{usage-indented}` replaces the hardwired indentation that only worked for
single-line usage strings.

The documentation for `App::override_usage()` is updated to describe how
to use multiline usage strings along with an example.

COMPATIBILITY

This does not change any API's, so there are no code compatibility
concerns. So the main concern is whether usage is displayed differently
in any scenarios. Special care was taken to cover various scenarios,
including cases where the user may have worked-around this issue (#1068) by
embedding their own indentation (with either tabs or spaces) into their
override usage string. The following scenarios were considered and are
believed to produce compatible usage/help outputs:

[x] User does not use App::override_usage()
[x] User provides a single-line usage string to App::override_usage()
[x] User uses a custom help template with legacy `{usage}` tag
[x] User provides App::override_usage() a multi-line usage string WITH
    their own indentation AND uses the default help templates (which now
    use the `{usage-indented}` tag.

The one scenario that would result in different output is a case where
clap itself generates a multi-line usage string AND the a user-defined
help template uses the `{usage}` tag. In this scenario, all but the first
usage line would be indented differently (not indented). The behavior for
this case is intentional because the benefits outweigh the incompatibility.
The benefit is that clap-generated multi-line usage strings no longer have
inconsistent, baked-in indentation which means that using the `{usage}` tag
in templates will be less surprising to users, especially considering that
`{usage-indented}` now exists. And the incompatibility is unlikely to be
encountered because it requires both using either SubcommandsNegateReqs or
ArgsNegateSubcommands and a custom template. And if a user encounters this
issue, the repair would be to replace `{usage}` with `{usage-indented}` in
their custom help template.

Signed-off-by: Peter Grayson <pete@jpgrayson.net>