Add support for ignoring errors on subcommands
Currently when setting `ignore_errors(true)` anywhere but the top level,
it would not have any effect. As a result this means it's not possible
to ignore errors for a subcommand while still enforcing correctness on
the top level.

This patch makes it possible to set `ignore_errors(true)` anywhere in
the nested subcommand chain and all children will ignore errors without
affecting the parent elements.

This is particularly helpful for external subcommands that are still
documented on the top level, making it possible to generically forward
all parameters while still documenting the externals subcommand on the
top level.