Improve error for `args_conflicts_with_subcommands`
Before this patch when running into errors due to
`args_conflicts_with_subcommands`, you'd get an output looking like
this:

```
error: The subcommand 'auth' wasn't recognized

	Did you mean 'auth'?
```

This is highly confusing and leaves the user without any information on
what actually went wrong.

This patch addresses this problem by introducing a new error message
dedicated for this scenario which will inform the user that the specific
combination of argument and subcommand used is not valid.